+++ packages/styles/src/packages/popover/index.ts
@@ -1,0 +1,17 @@
+import * as style from "./popover.style.css";
+import {
+  backgroundColorProperties,
+  paddingProperties,
+  colorProperties,
+  sprinkle,
+} from "./popover.sprinkle.css";
+
+export const popover = {
+  style,
+  sprinkle,
+  properties: {
+    backgroundColor: backgroundColorProperties,
+    padding: paddingProperties,
+    color: colorProperties,
+  },
+};