+++ packages/react/Tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Tooltip component allows us to display additional information in a non-intrusive way.
 
+## 2022-10-06
+
+### 💡 Others
+
+- Refactored tooltip package color scheme. ([#37](https://github.com/TiendaNube/nimbus-design-system/pull/37) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-10-01
 
 ### 📚 3rd party library updates